Updates passkeys feature in Test Harness to be one screen.
