+++ TestHarness/Application/Support/Entitlements/TestHarness.entitlements
@@ -2,19 +2,17 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-    <key>com.apple.developer.associated-domains</key>
-    <array>
-        <string>webcredentials:*.bitwarden.com</string>
-        <string>webcredentials:*.bitwarden.eu</string>
-        <string>webcredentials:*.bitwarden.pw</string>
-    </array>
-  <key>com.apple.security.application-groups</key>
+	<key>com.apple.developer.associated-domains</key>
+	<array>
+		<string>webcredentials:*.bitwarden.com</string>
+		<string>webcredentials:*.bitwarden.eu</string>
+		<string>webcredentials:*.bitwarden.pw</string>
+	</array>
+	<key>com.apple.security.application-groups</key>
 	<array>
 		<string>group.${BASE_BUNDLE_ID}</string>
 		<string>$(SHARED_APP_GROUP_IDENTIFIER)</string>
 	</array>
-	<key>com.apple.developer.authentication-services.autofill-credential-provider</key>
-	<true/>
 	<key>keychain-access-groups</key>
 	<array>
 		<string>$(AppIdentifierPrefix)com.bitwarden.testharness</string>
