--- TestHarnessShared/UI/Autofill/CreatePasskey/CreatePasskeyEffect.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-/// Effects that can be processed by a `CreatePasskeyProcessor`.
-///
-enum CreatePasskeyEffect: Equatable {
-    /// The user tapped the Register Passkey button.
-    case registerPasskey
-}
