--- TestHarnessShared/UI/Autofill/ManagePasskeys/ManagePasskeysAction.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-/// Actions that can be processed by a `ManagePasskeysProcessor`.
-///
-enum ManagePasskeysAction: Equatable {}
