--- TestHarnessShared/UI/Autofill/UsePasskey/UsePasskeyAction.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-/// Actions that can be processed by a `UsePasskeyProcessor`.
-///
-enum UsePasskeyAction: Equatable {
-    /// The relying party ID field was updated.
-    case rpIdChanged(String)
-}
