+++ TestHarnessShared/UI/Platform/Root/RootCoordinator.swift
@@ -37,20 +37,16 @@ class RootCoordinator: Coordinator, HasStackNavigator {
 
     func navigate(to route: RootRoute, context: AnyObject?) {
         switch route {
-        case .registerPasskey:
-            showCreatePasskey()
         case .cardAutofillForm:
             showCardAutofillForm()
         case .fileShare:
             showFileShare()
-        case .managePasskeys:
-            showManagePasskeys()
+        case .passkeys:
+            showPasskeys()
         case .scenarioPicker:
             showScenarioPicker()
         case .simpleLoginForm:
             showSimpleLoginForm()
-        case .usePasskey:
-            showUsePasskey()
         case .totpAutofillForm:
             showTOTPAutofillForm()
         }
@@ -62,27 +58,15 @@ class RootCoordinator: Coordinator, HasStackNavigator {
 
     // MARK: Private Methods
 
-    /// Shows the create passkey test screen.
+    /// Shows the unified passkey scenario screen.
     ///
-    private func showCreatePasskey() {
-        let processor = CreatePasskeyProcessor(
+    private func showPasskeys() {
+        let processor = PasskeyScenarioProcessor(
             coordinator: asAnyCoordinator(),
             delegate: self,
             passkeyRegistryService: services.passkeyRegistryService,
         )
-        let view = CreatePasskeyView(store: Store(processor: processor))
-        let viewController = UIHostingController(rootView: view)
-        stackNavigator?.push(viewController)
-    }
-
-    /// Shows the manage passkeys screen.
-    ///
-    private func showManagePasskeys() {
-        let processor = ManagePasskeysProcessor(
-            coordinator: asAnyCoordinator(),
-            passkeyRegistryService: services.passkeyRegistryService,
-        )
-        let view = ManagePasskeysView(store: Store(processor: processor))
+        let view = PasskeyScenarioView(store: Store(processor: processor))
         let viewController = UIHostingController(rootView: view)
         stackNavigator?.push(viewController)
     }
@@ -124,15 +108,6 @@ class RootCoordinator: Coordinator, HasStackNavigator {
         stackNavigator?.push(viewController)
     }
 
-    /// Shows the use passkey test screen.
-    ///
-    private func showUsePasskey() {
-        let processor = UsePasskeyProcessor(coordinator: asAnyCoordinator(), delegate: self)
-        let view = UsePasskeyView(store: Store(processor: processor))
-        let viewController = UIHostingController(rootView: view)
-        stackNavigator?.push(viewController)
-    }
-
     /// Shows the TOTP autofill form test screen.
     ///
     private func showTOTPAutofillForm() {
@@ -149,18 +124,14 @@ extension RootCoordinator: HasErrorAlertServices {
     var errorAlertServices: ErrorAlertServices { services }
 }
 
-// MARK: - CreatePasskeyProcessorDelegate
+// MARK: - PasskeyScenarioProcessorDelegate
 
-extension RootCoordinator: CreatePasskeyProcessorDelegate {
-    func presentationAnchorForPasskeyRegistration() async -> ASPresentationAnchor {
+extension RootCoordinator: PasskeyScenarioProcessorDelegate {
+    func presentationAnchorForPasskeyAssertion() async -> ASPresentationAnchor {
         stackNavigator?.rootViewController?.view.window ?? UIWindow()
     }
-}
 
-// MARK: - UsePasskeyProcessorDelegate
-
-extension RootCoordinator: UsePasskeyProcessorDelegate {
-    func presentationAnchorForPasskeyAssertion() async -> ASPresentationAnchor {
+    func presentationAnchorForPasskeyRegistration() async -> ASPresentationAnchor {
         stackNavigator?.rootViewController?.view.window ?? UIWindow()
     }
 }
